[#250] refactor: 타이머 패키지 구조 변경
