+++ linkmind/src/main/java/com/app/toaster/domain/Category.java
@@ -1,12 +1,10 @@
 package com.app.toaster.domain;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 
+import com.app.toaster.timer.domain.Reminder;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
