+++ linkmind/src/main/java/com/app/toaster/service/timer/TimerService.java
@@ -1,14 +1,14 @@
 package com.app.toaster.service.timer;
 
-import com.app.toaster.controller.request.timer.CreateTimerRequestDto;
-import com.app.toaster.controller.request.timer.UpdateTimerCommentDto;
-import com.app.toaster.controller.request.timer.UpdateTimerDateTimeDto;
-import com.app.toaster.controller.response.timer.CompletedTimerDto;
-import com.app.toaster.controller.response.timer.GetTimerPageResponseDto;
-import com.app.toaster.controller.response.timer.GetTimerResponseDto;
-import com.app.toaster.controller.response.timer.WaitingTimerDto;
+import com.app.toaster.timer.controller.request.CreateTimerRequestDto;
+import com.app.toaster.timer.controller.request.UpdateTimerCommentDto;
+import com.app.toaster.timer.controller.request.UpdateTimerDateTimeDto;
+import com.app.toaster.timer.controller.response.CompletedTimerDto;
+import com.app.toaster.timer.controller.response.GetTimerPageResponseDto;
+import com.app.toaster.timer.controller.response.GetTimerResponseDto;
+import com.app.toaster.timer.controller.response.WaitingTimerDto;
 import com.app.toaster.domain.Category;
-import com.app.toaster.domain.Reminder;
+import com.app.toaster.timer.domain.Reminder;
 import com.app.toaster.domain.User;
 import com.app.toaster.exception.Error;
 import com.app.toaster.exception.model.CustomException;
@@ -16,7 +16,7 @@ import com.app.toaster.exception.model.ForbiddenException;
 import com.app.toaster.exception.model.NotFoundException;
 import com.app.toaster.exception.model.UnauthorizedException;
 import com.app.toaster.infrastructure.CategoryRepository;
-import com.app.toaster.infrastructure.TimerRepository;
+import com.app.toaster.timer.infrastructure.TimerRepository;
 import com.app.toaster.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
