+++ linkmind/src/main/java/com/app/toaster/timer/controller/request/UpdateTimerCommentDto.java
@@ -1,4 +1,4 @@
-package com.app.toaster.controller.request.timer;
+package com.app.toaster.timer.controller.request;
 
 public record UpdateTimerCommentDto(String newComment) {
 }
