+++ linkmind/src/main/java/com/app/toaster/timer/controller/response/CompletedTimerDto.java
@@ -1,10 +1,6 @@
-package com.app.toaster.controller.response.timer;
+package com.app.toaster.timer.controller.response;
 
-import com.app.toaster.controller.response.search.CategoryResult;
-import com.app.toaster.controller.response.search.SearchCategoryResult;
-import com.app.toaster.domain.Reminder;
-
-import java.util.List;
+import com.app.toaster.timer.domain.Reminder;
 
 public record CompletedTimerDto(Long timerId, Long categoryId, String remindTime, String remindDate, String comment) {
     public static CompletedTimerDto of(Reminder timer,String remindTime, String remindDate){
