+++ linkmind/src/main/java/com/app/toaster/timer/controller/response/WaitingTimerDto.java
@@ -1,10 +1,8 @@
-package com.app.toaster.controller.response.timer;
+package com.app.toaster.timer.controller.response;
 
-import com.app.toaster.domain.Reminder;
+import com.app.toaster.timer.domain.Reminder;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 public record WaitingTimerDto(Long timerId, String remindTime, String remindDates, Boolean isAlarm, LocalDateTime updateAt, String comment, Long categoryId) {
     public static WaitingTimerDto of(Reminder timer, String remindTime, String remindDates) {
