+++ linkmind/src/main/java/com/app/toaster/timer/domain/Reminder.java
@@ -1,5 +1,8 @@
-package com.app.toaster.domain;
+package com.app.toaster.timer.domain;
 
+import com.app.toaster.domain.Category;
+import com.app.toaster.domain.IntegerListConverter;
+import com.app.toaster.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
